# backend/specs/001-rag-chatbot/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: RAG Chatbot API
  version: 1.0.0
  description: API for retrieving grounded answers from the book.
paths:
  /ingest:
    post:
      summary: Trigger Content Ingestion
      operationId: trigger_ingestion
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                source_dir:
                  type: string
                  description: "Path to docs directory (optional, defaults to env config)"
                  example: "docs/"
                force_reindex:
                  type: boolean
                  default: false
      responses:
        '202':
          description: Ingestion started
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: integer
                  status:
                    type: string
                    example: "pending"
        '401':
          description: Unauthorized
        '429':
          description: Too Many Requests

  /query:
    post:
      summary: Ask a Question
      operationId: ask_question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '429':
          description: Rate limit exceeded
        '500':
          description: Internal Server Error

  /admin/reindex:
    post:
      summary: Force Re-index (Admin)
      operationId: force_reindex
      security:
        - bearerAuth: []
      responses:
        '202':
          description: Re-index started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '403':
          description: Forbidden (Admin only)

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    QueryRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          example: "What is an Agentic Workflow?"
        selection_context:
          type: string
          description: "User-selected text from the book"
        selection_only:
          type: boolean
          default: false
          description: "If true, answer only using selection_context"

    QueryResponse:
      type: object
      properties:
        answer:
          type: string
          description: "The generated answer (Markdown)"
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
        provenance:
          type: object
          properties:
            mode:
              type: string
              enum: [global, selection]
            latency_ms:
              type: number

    Source:
      type: object
      properties:
        title:
          type: string
          example: "Chapter 1: Introduction"
        url:
          type: string
          example: "/docs/intro#section-1"
        score:
          type: number
          example: 0.89
        snippet:
          type: string

    JobResponse:
      type: object
      properties:
        job_id:
          type: integer
        status:
          type: string